+++ okitweb/static/okit/model/js/artefacts/block_storage_volume.js
@@ -23,11 +23,6 @@ class BlockStorageVolume extends OkitArtifact {
         // Update with any passed data
         this.merge(data);
         this.convert();
-        // Check if built from a query
-        if (this.availability_domain.length > 1) {
-            this.region_availability_domain = this.availability_domain;
-            this.availability_domain = this.region_availability_domain.slice(-1);
-        }
     }