+++ okitweb/static/okit/panels/js/base_panel.js
@@ -2,7 +2,10 @@
 ** Copyright (c) 2020, 2021, Oracle and/or its affiliates.
 ** Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 */
-console.info('Loaded Views Javascript');
+console.info('Loaded OkitPanel Javascript');
 
-// const okitViewClasses = [];
-// const okitViews = [];
+/*
+** Define OkitPanel Class
+*/
+class OkitPanel {
+}